os/bluestore: BufferSpace per Onode
Instead of indexing buffers by blob offset we do it by logical offset.
Previously we tried using physical offsets but it was a pretty bad idea
is it required extra bound checks of physical extents and extra loops
for to add them, that includes chopping off buffer data.

Logical offsets don't have these problems but we have different ones.
Bufferlist data is padded before being used in `_do_alloc_write` which
means before adding it to BufferSpace, we are required to get the
unpadded substr_of data because padded data might overflow to an invalid
logical offset in contrast to when we used blob offsets where padded
data was wonderful. This is true in in `generate_read_result_bl` too, we
read by regions of blob extents but we must translate it back to logical
offsets, that means an extra loop to include correct logical regions was

Signed-off-by: Pere Diaz Bou <pere-altea@hotmail.com>