qa: Add benign cluster warning from ec-inconsistent-hinfo test to ignorelist
The changes introduced in PR: https://github.com/ceph/ceph/pull/53524
made the randomized values of osd_op_queue and osd_op_queue_cut_off
consistent across all OSD shards.

Due to the above, ec-inconsistent-hinfo test could fail with the following
cluster warning (benign) depending on the randomly selected scheduler type.

"cluster [WRN] Error(s) ignored for 2:ad551702:::test:head enough copies
available"

In summary, the warning is generated due to the difference in the PG
deletion rates between WPQ and mClock schedulers. Therefore, the warning
shows up in cases where the mClock scheduler is the op queue scheduler
chosen randomly for the test. The PG deletion rate with mClock scheduler
is quicker compared to the WPQ scheduler since it doesn't use sleeps
between each delete transaction and relies on the cost of the deletion
which in turn is proportional to the average size of the objects in the PG.

For a more detailed analysis, see the associated tracker.

Fixes: https://tracker.ceph.com/issues/64573
Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>