mon/OSDMonitor: always create pgs in the epoch in which the pool was created
The logic here didn't actually work.  If update_pending_pgs doesn't get
to a particular PG (because of mon_osd_max_creating_pgs limitation) in
the epoch in which the pool was created, the pg will end up with a
same_interval_since at that epoch rather than the pool creation epoch.
This can cause an IO submitted by a client based on the epoch in which
the pool was created to be rejected by the OSD without an interval
change actually occuring and therefore without the client resending the
op.

In order to make this limit actually function, once we actually process
a pending pg, we'd have to go back to the OSDMap at which the pool was
created and work forward to get the correct interval bound.  That seems
even more expensive, so instead this patch simply removes the limit.

Fixes: https://tracker.ceph.com/issues/64546
Signed-off-by: Samuel Just <sjust@redhat.com>