cephadm: reorder tox env to run fastest tasks first
Reorder the tox environments so that the simplest, fastest running tools
execute first. This can be handy when running tox locally (say in a
pre-commit workflow) and you get to see a formatting or mypy error
before the slower unit tests begin you can abort the tests and fix up
the formatting instead of waiting for all unit tests to run only to
find out you need to make changes to fix formatting problems.

Signed-off-by: John Mulligan <jmulligan@redhat.com>