mgr/object_format: allow easily disabling dict key sorting
When adapating an object to JSON or YAML formatting the previous version
of the code always sorted dict keys. This is nice for code that placed
no ordering on the dict keys. However, for newer code that may rely on
python's sorted dicts we want to be able to disable this behavior.
Add two new keyword args to the adapter type that allow disabling
the sorting.

Signed-off-by: John Mulligan <jmulligan@redhat.com>