mgr/cephadm: Fix unfound progress events
While applying service specs, cephadm creates a progress event for
the daemons to be added or deleted from the hosts. The progress
event is initialized if progress_total is greater than 0
but at the end cephadm tries to mark the progress event as
complete/fail even though progress event has not been initialized.

Mark progress events as complete/fail only if they are initialized.

Fixes: https://tracker.ceph.com/issues/65799

Signed-off-by: Prashant D <pdhange@redhat.com>