pybind/mgr: disable flake8 on a few non-obvious issues
Use the `noqa` comment on a handful of lines that flake8 detects as
issues. I don't want to invest a ton of time looking into if the
imports cause needed side-effects. I also don't want to debug if
the bare except is trying to catch exceptions that don't inherit
from Exception. So, make the file flake8 clean by just suppressing
these remaining things.

Signed-off-by: John Mulligan <jmulligan@redhat.com>