osdc/Striper: relax assert in assemble_result() for raw buffers
The previous commit uncovered an old bug, introduced in commit
f354a56da346 ("Striper: Add function 'assemble_result(CephContext *cct,
char *buffer, size_t len)' in StripedReadResult.").  The buffer is
allowed to be NULL when assemble_result() is called -- it becomes
a problem only if some data was read.

Strengthen ZeroLengthRead and ZeroLengthWrite tests to cover both
a bogus pointer and a NULL pointer for buffer.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>