.github: expand tests label to all files under qa
The test job definition under qa/suites is an integral part of almost
any test.  Often, the test logic is split between the task or workunit
and respective snippet(s) under qa/suites.

Other files under qa are less used, but still related to nothing but
testing, so just add the label on all of it.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>