rgw/period: move realm notify out of period commit
require the callers to trigger the realm notify manually.
RGWOp_Period_Post takes advantage of this to delay the realm notify
until after it's sent its response to the client. this avoids racing
with realm reload which will close the connection

commit returns EEXIST errors directly so the caller can avoid the
realm notify

Signed-off-by: Casey Bodley <cbodley@redhat.com>