common: MemoryModel: keep /proc files open
Allowing clients to hold on to the MemoryModel object for multiple
"snapshots", saving (per measurements) ~40% of the sampling run time.

Also: separating the parsing of /proc/maps (which consumes most of
the remaining execution time) into a separate function. This will allow
us to create a "no need for precise heap" fast mode (in a future
commit).

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>