mgr/smb: Use the default for minimum protocol version
'server min protocol' smb.conf option indicates the minimum SMB protocol
version that the server can support during client negotiation phase. We
can be generous in this case to accept various client implementations
with a subset of supported protocol versions. With 'SMB2' it selects the
highest available version from the list of sub protocols. Instead we
refrain from configuring the option to support the least possible SMB2
version by default(which is SMB2_02 at this point in time) from Samba.

ref: https://www.samba.org/samba/docs/current/man-html/smb.conf.5.html
Signed-off-by: Anoop C S <anoopcs@cryptolab.net>