Revert "test: reproduce flap peering bug (wip)"
This reverts commit ab678a17696e2a17d2554c2ecab7dd10743e718e.

Note from Dan van der Ster:
This is a test that should succeed,
it definitely used to succeed back in the L/O days of Ceph. At some
point peering code changed and this behaviour regressed. In short,
an OSD goes down then comes up, and no objects were modified in the mean time.
There should be no degraded PGs in this case.

As this commit is currently breaking make check on all PRs, I think it should
be re-evaluated and merged so whatever fix is needed along with this test to
make it work are merged together.

Fixes: https://tracker.ceph.com/issues/66556
Signed-off-by: Laura Flores <lflores@ibm.com>