crimson/osd/pg: run peering_state.handle_event without interruptor::async
Partial revert of bc23cf8.

With the removal of rebuild_missing_set_with_deletes_crimson, this
shouldn't be necessary.

In a larger sense, it's problematic to yield in the middle of handling
a peering event -- particularly AdvMap/ActMap -- as it allows other
continuations to observe PeeringState in the middle of handling an
event, which the external interface really does not support.  classic
disallows this by holding the pg lock while processing basically anything,
including peering events.  With crimson, it should suffice to simply
not yield while processing the event.

Signed-off-by: Samuel Just <sjust@redhat.com>