test: Create ParallelPGMapper object before start threadpool
In the case of creating a ParallelPGMapper object after
starting the threadpool the test case might crash.
The test case is crash while worker thread is trying to
delete a job but the object is still in the construction phase
not fully constructed.

Solution: To avoid a crash create the object before start
          a threadpool.

Fixes: https://tracker.ceph.com/issues/42884
Signed-off-by: Mohit Agrawal <moagrawa@redhat.com>