rgw: donot check for NotPrincipal in IsPublicStatement
As NotPrincipal cannot be used with Effect Allow there is no need to
check for that.
Also before this change, policies without `NotPrincipal` were identified
as a public as std::none_of() returns true on empty flats and the check
was not considering the fact that the policy might have pricipals. but
due to the missing concept from AWS about how to deal with Principal and
NotPrincipal on Allow Effect we can just remove it.

Signed-off-by: Seena Fallah <seenafallah@gmail.com>