cephadm: run smbmetrics as side-car container
When samba metrics-exporter image is provided and it is not an empty
string, run the smbmetrics as side-car container along-side main
samba-server. This metrics-exporter uses 'smbstatus' to convert SMB
stats into standard Prometheus metrics (via port 9922 by default).

This commit defines ContainerCommon as base of SambaContainerCommon and
SMBMetricsContainer. It defines the basic APIs of smb containers with
empty defaults. It also allows passing explicit image URL to samba's
sidecar containers. If no alternative image is provided (which is the
typical case), use the one provided by context or default one.

Signed-off-by: Shachar Sharon <ssharon@redhat.com>