debian pkg: record python3-packaging dependency for ceph-volume
Since commit 43d5f4b9687 ("ceph-volume: use 'no workqueue' options
with dmcrypt") the python "packaging" module is used to parse the
cryptsetup version output, but the debian packaging was not updated to
record that new dependency.

So simply record this in the d/control file, adding a <pkg>.requires
file seemed to not really winning us anything here.

Fixes: https://tracker.ceph.com/issues/67290
Fixes: 43d5f4b96875b0497fd0ce5f7949f3f0d83610d4
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>