crimson/osd: InternalClientRequests should not be repeated after interval change
Normal cient requests are special in that we "requeue" them after an
interval change if the primary did not change.  This behavior exists due
to a client-side optimization where the client doesn't resend the
operation unless the primary changed.

That's not true for InternalClientRequest (generally used for watch
expirations).  The primary will eventually reload and re-expire the
watch once the other watchers reconnect.

If we actually did want InternalClientRequests to be able to repeat
after interval change, they'd need the same resettable handle mechanism
normal ClientRequests use -- this implementation wouldn't actually work
as it would mean the same handle reentering a previous stage.

Fixes: https://tracker.ceph.com/issues/68068
Signed-off-by: Samuel Just <sjust@redhat.com>