rgw: remove rgw_data_log_obj_prefix
The evaluation based on rgw_data_log_obj_prefix had a logical error
not considering the value and always were returning "data_log".
As the config was having the dev level and apparantly it was useless
and fixing the logic could break the existing clusters (if they don't
adopt before upgrade), it was decided to remove the config.

Fixes: https://tracker.ceph.com/issues/67861
Signed-off-by: Seena Fallah <seenafallah@gmail.com>