librbd: introduce rbd_group_snap_namespace_type_t enum
Commit 2ffacc7f2141 ("rbd: add group snap info command") added
a commented out field in rbd_group_snap_info2_t struct but didn't
define the corresponding enum, expecting it to be brought in with the
patchset that introduces mirror group snapshots.  That work is still
in progress, but there is already interest in the new command and APIs
that use rbd_group_snap_info2_t struct and even a backport request.

Finalize the struct definition to avoid a breaking change for those
users in the future.  RBD_GROUP_SNAP_NAMESPACE_TYPE_USER can be treated
as dummy in terms of meaning until RBD_GROUP_SNAP_NAMESPACE_TYPE_MIRROR
is added.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>