rgw: use effective owner in PutBucketReplication
As PutBucketReplication requests are forwarded to the master, it
needs to use the effective owner to set the correct (client) as the
uid for sync pipe rather than the system user.

sync policies require a valid UID for authorization when operating in
rgw_sync_pipe_params::Mode::MODE_USER mode. Currently, when forwarding
requests to the master, rgwx-uid holds the ACLOwner string rather than
a UID, which can't be used for sync policy checks. Until this is
properly implemented, we are rejecting PutBucketReplication calls for
account holders.

Fixes: https://tracker.ceph.com/issues/68172
Signed-off-by: Seena Fallah <seenafallah@gmail.com>