doc: Document the Windows CI job
We're adding a document that describes:

* the steps performed by the Windows CI job
* artifact structure
* frequently asked questions

The document is meant to assist the Ceph developers in investigating
CI failures. This is especially important as the Windows CI job runs
integration tests that would otherwise only be executed by
Teuthology, thus helping catch potential regressions quickly.

Note that the identified regressions are not necessarily Windows
specific, usually affecting Linux builds as well.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>