doc/governance: revert abe9cf8ec46953d3cc05db9548765e63c6a68860
Revert the commit (manually, by restoring the file by hand to the state
it was in prior to abe9cf8) to the state that it was in before I added
the Executive Council Responsibilities document to governance.rst. This
document cannot be edited at will, but must be voted on by the
Leadership Team.

Signed-off-by: Zac Dover <zac.dover@proton.me>