common/Finisher: wake up after pushing to the queue
Pushing to the queue may take a long time when the `std::vector` needs
to allocate more memory.  We should wake up the `Finisher` thread only
right before unlocking the `finisher_mutex` to reduce lock contention,
because it is the more likely that the mutex can really be acquired
when the thread really wakes up.

This imitates how commit d6f42ced805cb2913 did it - it refactored only
one of the `queue()` overloads, leaving less-than-optimal copies of
this piece code in all other overloads.

Signed-off-by: Max Kellermann <max.kellermann@ionos.com>