mgr/dashboard: fix total objects/Avg object size in RGW Overview Page
Till now we are calculating the total number of objects and the average
object size in the RGW Overview Page using `ceph df` command's output.
As per the discussion with RGW team, this data is not correct as S3
objects in rgw can occupy more than one rados object. This PR tends to
make the overview page's info in sync with the RGW bucket page's info.

Fixes: https://tracker.ceph.com/issues/68733

Signed-off-by: Aashish Sharma <aasharma@redhat.com>