CODEOWNERS: let orchestration cover all of python-common
While it's not perfectly correct to say the orchestration team owns
src/python-common, it's closer to the truth than not. It also ensures
that someone with domain specific knowledge is looking at the rest of
python-common outside just the deployment section (short of creating
some sort of new python-sig type group). This is relevant due to the
changes in PR #60377 as python-common's ceph package will be included
in cephadm.

Signed-off-by: John Mulligan <jmulligan@redhat.com>