python-common: use tox.ini envlist defaults when run via cmake
Match recent(ish) changes to CMakeLists.txt for src/pybind/mgr and
stop hard-coding a list of tox envs to run when testing is triggered
via cmake.
Instead, automatically reuse the defaults in tox.ini. This is good
because we recently changed those defaults and this change makes
'make check' run the same set of tests and checks again.

Signed-off-by: John Mulligan <jmulligan@redhat.com>