etc/sysctl: adjust fs.aio-max-nr for crimson
This will switch 90-ceph-osd.conf when compiling Crimson builds.

See e5a080c.
Clasical osd packages are adjusting the aio-max-nr value
from the default of 65536 to 1048576 as the default is too low
for some deployments. The same should be applied for Crimson's
packages.

No further changes are applied to ceph.spec.in / debuan/rules since
etc/sysctl/90-ceph-osd.conf will be installed regardless.
For rpm, we package crimson-osd with ceph-osd - this means that
the existing `%post osd` will be applied for crimson-osd packages
as well.

Signed-off-by: Matan Breizman <mbreizma@redhat.com>