osd/scrub: abort reserving scrub if an operator-initiated scrub is requested
Handling the case of receiving an operator command while the PG is
scrubbing, but
is waiting for replicas' reservations:

Now that the reservations are queued, the wait may be a very prolonged
one.
Usually - an operator direct scrub command has a priority high enough
to not require waiting for reservations. But in the current
implementation,
it would wait until the running scrub session terminates, and only then
will rerun at that high priority. This is not the intended behavior.

The solution is to abort the existing scrub session, and start the new
one.

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>