osd: implement ECSwitch to switch between EC implementations
We have introduced a vaneer class to wrap the two implementations of EC.

The intent here is to isolate the legacy EC code as much as possible, meaning that it can be left alone and therefore the regression risk can be reduced.

We might choose to reverse this and have two different classes extending ECBackend.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>

FIXUP implement EC Switch

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>