qa/cephadm: wait between starting and restarting osd in test_daemon_restart test
The test before the pause test ends with `ceph orch daemon restart osd.0`
which may temporarily cause a CEPHADM_FAILED_DAEMON warning for that
OSD. It seems to be failing with "status=219/CGROUP", logging nothing
else of note, and coming up totally fine when restarted by systemd
a second later (see the linked tracker for more info/logs). This
sleep is me thinking this may be timing related and that if we
separate the start from the restart by a few more seconds it might
avoid the error.

Fixes: https://tracker.ceph.com/issues/69526

Signed-off-by: Adam King <adking@redhat.com>