doc/cephfs: remove link to undeveloped tool
Remove a link to a tracker for a planned-but-never-developed tool that
would have facilitated the importation or exportation of a RADOS pool to
a file. This is a part of a project to improve the CephFS Disaster
Recovery documentation, which was requested by Eugen Block. See the
tracker issue below for more on this project.

Follows https://github.com/ceph/ceph/pull/61411

See https://github.com/ceph/ceph/pull/61411#discussion_r1918525109

https://tracker.ceph.com/issues/69557

Signed-off-by: Zac Dover <zac.dover@proton.me>