mgr/cephadm: add Server_Scope = <fsid> to NFSv4 section of ganesha conf
From the ganesha team

"""
In the NFSv4 param block, we need a parameter Server_Scope set to some value common among all servers in a cluster.

The default with it blank is to use the hostname which may be different for each server in the cluster.
"""

This is related to ongoing work on high availability nfs. From the cephadm side
we just need to make sure all nfs daemons in the cluster end up with
the same value for the Server_Scope field. This patch uses the cluster
id (which we already brought into the template as the "namespace" attribute)

Signed-off-by: Adam King <adking@redhat.com>