qa/standalone/scrub: retry in TEST_abort_periodic_for_operator()
The scenario created in the test requires two PGs that share the same
Primary, and at least one more member of their active sets.
The PR handles the (unlikely) case where the first PG tried does not
have a counterpart that matches the requirements.

A minor related fix in scrub-helpers.sh: using 'printf' to
return a value, instead of an out (ref) parameter.

Fixes: https://tracker.ceph.com/issues/69837
Signed-off-by: Ronen Friedman <rfriedma@redhat.com>