mds: Store remote inode number in referent inode
Hardlink referent inode plumbing continues.
The remote inode information contains the following.
  1. remote_ino - remote inode number
  2. d_type
  3. alternate_name

With the introduction of the referent inode as a full
blown CInode for the hardlink, these information needs
to be part of the inode. Hence add 'remote_ino' field
to the inode. The exisiting 'd_type' and 'alternate_name'
fields could be used.

Fixes: https://tracker.ceph.com/issues/54205
Signed-off-by: Kotresh HR <khiremat@redhat.com>