mgr: add status command
The Rook operator relies on a `status` (asok) command to the mgr to verify it
is operational. However, this command was actually served by the CephFS
`Client` library that the ceph-mgr daemon statically linked in and used for
sending MDS commands. That statically linked client was removed via f7ff58c6
(case insensitive directory tree feature).

So, add a legitimate ceph-mgr status command which simply outputs an empty
dictionary and returns success (0). TBD on adding useful information.

Fixes: f7ff58c6e8aff1b9e4735e3c22c4dbd4fc6afb92
Fixes: https://tracker.ceph.com/issues/70571
Signed-off-by: Patrick Donnelly <pdonnell@ibm.com>