osd/PeeringState: Introduce is_missing_any_head_or_clone_of
When checking if a client requested object is missing, we must first
obtain the actual (resolved) object from the OSD prespective. See how
Crimson handles this logic with resolve_oid (Classic does the same).
For us to resolve an object we *must* obtain its object context.

There are some scenarios (e.g serving replicated reads) where checking
if the object (head or clone) is missing prior to obtaining the obc
could be useful to avoid further lookups or unnecessarily complicated/
bug-prone pathways.

With this function, we can check if any head or clone of this object
is missing, without resolving the (client) requested object.

Signed-off-by: Matan Breizman <mbreizma@redhat.com>