buildcontainer-setup.sh: Fix curl package conflict
It looks like at some point the centos9 image started shipping with
curl-minimal, which conflicts with the regular curl package. Asking dnf to find
the binary avoids this, since both packages provide it. Since we were already
doing this with rpmbuild, we can go ahead and loop wget into that in case
something similar happens there.

Signed-off-by: Zack Cerza <zack@cerza.org>