mgr/cephadm: do not delete smb fs cephx keys
This change effectively disables fencing for the smb service because
the previous attempt to implement fencing would destroy the only
cephx key. Deleting this key would prevent any smb service part of
the logical cluster from talking to cephfs, even ones that were not
to be fenced.

The whole concept of fencing and ranks needs a bit of a rethink in
regards to smb. For now, we're just going to rely on ctdb and not
cephadm for smb's HA.

Fixes: af0ebab0a7f0efd4fa48a5b4c0d1acbe1f8883b1
Signed-off-by: John Mulligan <jmulligan@redhat.com>