doc/radosgw: Fix RST syntax rendeded as text in oidc.rst
Empty line after starting a pre-formatted block with the double-colon
syntax is required, otherwise the double-colon does nothing and is just
rendered as-is as "::" and there would be no following pre-formatted
block.

Add empty lines after the double-colon syntax so that the following
block is rendered pre-formatted.

Also add bash privileged prompts to a block with 2 example CLI commands.

Signed-off-by: Ville Ojamo <14869000+bluikko@users.noreply.github.com>