doc/radosgw: Improve formatting in bucketpolicy.rst
Add a missing empty line before an unordered list that
caused it to be rendered as just normal text in a
single paragraph, instead of a list.

Use inline code for CLI command names and other
such identifiers.

Capitalize "s3" correctly as "S3" when it refers to
the S3 service/API.

Signed-off-by: Ville Ojamo <14869000+bluikko@users.noreply.github.com>