dashboard: fix the workaround for unpacking node sources
My previous workaround in the dashboard for the unpacking of non-root
own tarball as the fake root of a container did not work because of the
strange quoting/escaping behavior of cmake (it tried to run `id -u` as a
single command, not a command and an argument).
Use single quoted string and old school backticks to work around this issue.

Fixes: d6b4f8465a0e25b41be634d93136bc9de3bd7014
Signed-off-by: John Mulligan <jmulligan@redhat.com>