crimson/os/seastore: increase OI_MAX_LENGTH to avoid storing OI data in the omap tree
Due to a recent change (adding shard_versions in object_info_t),
the size of OI has increased to 236. This causes to store OI data to the omap tree
because the size is over than OI_MAX_LENGTH. As a result, this results in
performance drop in small random write.

Signed-off-by: Myoungwon Oh <ohmyoungwon@gmail.com>