libcephfs: Bump API major version
We recently had ABI changes[1] with respect to APIs from chown() family
which calls for a change in major version. Native users of the library
may not have to change their code but expected sizes differ when data
type for parameters are changed. However go-ceph, Go bindings for ceph,
couldn't build[2] unless the ABI change is made visible to the consumers
of the API. Following the Semantic Versioning guidelines[3] we reset
minor and patch (extra) versions to 0.

[1] https://github.com/ceph/ceph/pull/62934
[2] https://github.com/ceph/go-ceph/issues/1119
[3] https://semver.org/#spec-item-8

Fixes: https://tracker.ceph.com/issues/71519

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>